bug fix and add documentation
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@1034 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlTools.cpp
@@ -292,7 +292,7 @@ namespace tlp {
     if(!inScreen){
       return -1;
     }else{
-      return sqrt((bbBox[2]-bbBox[0])*(bbBox[2]-bbBox[0])+(bbBox[3]-bbBox[1])*(bbBox[3]-bbBox[1])) * 2;
+      return sqrt((double)(bbBox[2]-bbBox[0])*(double)(bbBox[2]-bbBox[0])+(double)(bbBox[3]-bbBox[1])*(double)(bbBox[3]-bbBox[1])) * 2;
     }
   }
   //====================================================
